requirements.txt
PyPDF2
openai
python-dotenv

Flask
flask-cors

azure-search-documents